Calculator can now use floating point numbers

+++ src/operations.rs
@@ -1,15 +1,15 @@
-pub fn add(n1: i32, n2: i32) -> i32 {
+pub fn add(n1: f64, n2: f64) -> f64 {
     n1 + n2
 }
 
-pub fn sub(n1: i32, n2: i32) -> i32 {
+pub fn sub(n1: f64, n2: f64) -> f64 {
     n1 - n2
 }
 
-pub fn mult(n1: i32, n2: i32) -> i32 {
+pub fn mult(n1: f64, n2: f64) -> f64 {
     n1 * n2
 }
 
-pub fn div(n1: i32, n2: i32) -> i32 {
+pub fn div(n1: f64, n2: f64) -> f64 {
     n1 / n2
 }
